rename exercise dl_dig_test to basic_computer_vision_dl
