+++ exercises/static/exercises/basic_computer_vision_dl/react-components/BasicComputerVisionRR.js
@@ -1,0 +1,14 @@
+import * as React from "react";
+import {Fragment} from "react";
+
+import "./css/BasicComputerVisionRR.css";
+
+const BasicComputerVisionRR = (props) => {
+  return (
+    <Fragment>
+    {props.children}
+    </Fragment>
+  );
+};
+
+export default BasicComputerVisionRR;
