+++ exercises/static/exercises/basic_computer_vision_dl/react-components/css/BasicComputerVisionRR.css
@@ -1,0 +1,28 @@
+* {
+    box-sizing: border-box;
+}
+
+body, html {
+    width: 100%; height: 100%;
+}
+
+#exercise-container {
+    position: absolute;
+    top: 0; left: 0; bottom: 0; right: 0;
+    overflow: hidden;
+    display: flex;
+    flex-direction: column;
+}
+
+#exercise-container #content {
+    width: 100%;
+    height: 100%;
+    overflow: hidden;
+}
+
+#exercise-container #content #content-exercise {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    height: 100%;
+}
