+++ exercises/static/exercises/basic_computer_vision_dl/react-components/css/GUICanvas.css
@@ -1,0 +1,5 @@
+.image {
+  position: absolute;
+  width: fit-content;
+  height: 100%;
+}
