--- exercises/static/exercises/dl_dig_test/react-components/BasicComputerVisionRR.js
@@ -1,14 +1,0 @@
-import * as React from "react";
-import {Fragment} from "react";
-
-import "./css/BasicComputerVisionRR.css";
-
-const BasicComputerVisionRR = (props) => {
-  return (
-    <Fragment>
-    {props.children}
-    </Fragment>
-  );
-};
-
-export default BasicComputerVisionRR;
