--- exercises/static/exercises/dl_dig_test/react-components/css/GUICanvas.css
@@ -1,5 +1,0 @@
-.image {
-  position: absolute;
-  width: fit-content;
-  height: 100%;
-}
