--- exercises/static/exercises/dl_dig_test/test_folder/test_file.txt
@@ -1,1 +1,0 @@
-hello there
