--- exercises/templates/exercises/dl_dig_test/exercise.html
@@ -1,59 +1,0 @@
-{% extends "react_frontend/exercise_base.html" %}
-{% load react_component %}
-
-{% block exercise_header %}
-{% endblock %}
-
-{% block react-content %}
-    {% if deployment %}
-        {% react_component exercise/rescue_people/BasicComputerVisionRR %}
-            {% react_component components/containers/MaterialBox id="exercise-container" %}
-            {% react_component components/common/MainAppBar exerciseName="Basic Computer Vision" url="https://jderobot.github.io/RoboticsAcademy/exercises/Drones/basic_computer_vision"%} 
-            {% end_react_component %}
-            {% react_component components/containers/MaterialBox id="content" %}
-                {% react_component components/containers/MaterialBox id="content-exercise" %}
-                    {% react_component components/containers/ExerciseControl %}{% end_react_component %}
-                    {% react_component components/containers/FlexContainer row %}
-                    {% react_component components/containers/FlexContainer console %} 
-                    {% react_component components/editors/EditorRobot %}{% end_react_component %}
-                        {% react_component components/visualizers/VncConsoleViewer%}{% end_react_component %}
-                        {% end_react_component%}
-                        {% react_component components/containers/FlexContainer%} 
-                            {% react_component exercise/rescue_people/SpecificBasicComputerVision %} {% end_react_component %}
-                            {% react_component components/visualizers/Camera%}{% end_react_component %}
-                        {% end_react_component %}
-                {% end_react_component %}
-            {% end_react_component %}
-            {% react_component components/message-system/Loading %}{% end_react_component %}
-            {% react_component components/message-system/Alert %}{% end_react_component %}
-{% end_react_component %}
-{% end_react_component %}
-    {% end_react_component %}
-    {% else %}
-    {% react_component exercise/basic_computer_vision/BasicComputerVisionRR %}
-        {% react_component components/wrappers/MaterialBox id="exercise-container" %}
-            {% react_component components/layout_components/MainAppBar exerciseName="Basic Computer Vision" url="https://jderobot.github.io/RoboticsAcademy/exercises/ComputerVision/basic_computer_vision"%} 
-            {% end_react_component %}  
-            {% react_component components/wrappers/MaterialBox id="content" %}
-                {% react_component components/wrappers/MaterialBox id="content-exercise" %}
-                    {% react_component components/layout_components/ExerciseControl %}{% end_react_component %}
-                    {% react_component components/wrappers/FlexContainer row %}
-                    {% react_component components/wrappers/FlexContainer console %} 
-                        {% react_component components/editors/EditorRobot %}{% end_react_component %}
-                        {% react_component components/visualizers/ConsoleViewer%}{% end_react_component %}
-                        {% end_react_component%}
-                        {% react_component components/wrappers/FlexContainer%} 
-                            {% react_component exercise/basic_computer_vision/SpecificBasicComputerVision %}{% end_react_component %}
-                            {% react_component exercise/dl_dig_test/FileUploadComponent %}{% end_react_component %}
-                            {% react_component components/visualizers/Camera%}{% end_react_component %}
-                        {% end_react_component %}
-                {% end_react_component %}
-            {% end_react_component %}
-        {% react_component components/message_system/Loading %}{% end_react_component %}
-        {% react_component components/message_system/Alert %}{% end_react_component %}
-{% end_react_component %}
-{% end_react_component %}
-    {% end_react_component %}
-    {% endif %}
-{% endblock %}
-
